+++ packages/chart/src/common/config.ts
@@ -2,7 +2,7 @@ export const DEFAULT_VALS = {
   MARGIN: {
     top: 20,
     right: 20,
-    bottom: 30,
-    left: 60,
+    bottom: 50,
+    left: 55,
   },
 };