+++ packages/chart/src/hooks/useChartDimensions.ts
@@ -28,8 +28,10 @@ export const useChartDimensions = (
   // compute the outer and inner dimension of the chart
   const chartRef = useRef<HTMLDivElement>(null);
   const legendRef = useRef<HTMLDivElement>(null);
+  const titleRef = useRef<HTMLDivElement>(null);
   const outerDimension: GraphDimension = useContainerDimension(chartRef);
   const legendDimension = useContainerDimension(legendRef);
+  const headerDimension = useContainerDimension(titleRef);
   const legendOrient = (color && color.legend && color.legend.orient) || 'right';
 
   const { graphDimension, graphMargin }  = useMemo(
@@ -37,16 +39,26 @@ export const useChartDimensions = (
       return getInnerGraphDimensionAndMargin(
         outerDimension,
         margin,
+        headerDimension,
         legendDimension,
         legendOrient,
       );
     },
-    [outerDimension, margin, legendDimension, legendOrient],
+    [
+      outerDimension,
+      margin,
+      headerDimension,
+      legendDimension,
+      legendOrient,
+    ],
   );
 
   return {
     /** Ref to the chart, which is to be passed in the props of the container */
     chartRef,
+
+    /** Ref to the header box, which is to be passed in the props of the header box */
+    titleRef,
 
     /** Ref to the legend, which is to be passed in the props of the legend */
     legendRef,