+++ packages/graph/src/index.ts
@@ -1,4 +1,3 @@
-export * from './components/Foo';
 export * from './common/types';
 export * from './common/config';
 export * from './hooks/useContainerDimension';
@@ -8,6 +7,7 @@ export * from './layers/ResponsiveLayer';
 export * from './layers/AxisLayer';
 export * from './layers/HoverLayer';
 export * from './layers/TooltipLayer';
+export * from './overlays/HeaderBox';
 export * from './themes';
 export * from './legend/Legend';
 export * from './legend/LegendGroup';
@@ -19,3 +19,4 @@ export * from './utils/getColorScale';
 export * from './utils/getDataGroupByEncodings';
 export * from './utils/getRecordFieldSelector';
 export * from './utils/getFieldValuesFromData';
+export * from './utils/styled-components';