+++ packages/graph/src/layers/AxisLayer.tsx
@@ -1,7 +1,7 @@
-import React, { useContext } from 'react';
+import React, { useContext, useMemo } from 'react';
 import { AxisBottom, AxisLeft } from '@vx/axis';
 
-import { AxisScale } from '../common/types';
+import { AxisScale, AxisEncoding } from '../common/types';
 import { ThemeContext } from '../themes';
 
 /**
@@ -14,6 +14,18 @@ function getNumberOfTicks(axisLength: number, data: object[]): number {
   const lengthBasis = axisLength > 0 ? axisLength : 60;
   const ticksFromLen = Math.ceil(lengthBasis / 60);
   return Math.min(maxTicks, ticksFromLen);
+}
+
+/**
+ * Returns the text of the label displayed along axis
+ */
+function getAxisLabel(axis: AxisEncoding): string | undefined {
+  const { label, field, hideLabel } = axis;
+  if (hideLabel) {
+    return undefined;
+  }
+
+  return label || field;
 }
 
 export interface AxisLayerProps {
@@ -32,11 +44,17 @@ export interface AxisLayerProps {
   /** Data of the chart */
   data: object[];
 
+  /** Axis encoding of x-axis */
+  x: AxisEncoding;
+
+  /** Axis encoding of y-axis */
+  y: AxisEncoding;
+
   /** X-axis configurations produced by `getAxisScale` */
-  xAxis: AxisScale;
+  xAxisScale: AxisScale['scale'];
 
   /** Y-axis configurations produced by `getAxisScale` */
-  yAxis: AxisScale;
+  yAxisScale: AxisScale['scale'];
 }
 
 const getXtickLabelProps = (styles: {
@@ -67,32 +85,43 @@ const tickFormat = (val: any) => `${val}`;
 export const AxisLayer: React.SFC<AxisLayerProps> = ({
   width,
   height,
+  x,
+  y,
   // it should always show the left axis by default
   showLeftAxis = true,
   // it should always show the bottom axis by default
   showBottomAxis = true,
   data,
-  xAxis,
-  yAxis,
+  xAxisScale,
+  yAxisScale,
 }) => {
 
   const theme = useContext(ThemeContext);
   const { xAxis: xAxisTheme, yAxis: yAxisTheme } = theme;
 
+  const xAxisLabel = useMemo(() => getAxisLabel(x), [x]);
+  const yAxisLabel = useMemo(() => getAxisLabel(y), [y]);
+
   return (
       <>
+        {/* Y Axis */}
         {showLeftAxis && (
           <AxisLeft
             top={0}
             left={0}
-            scale={yAxis.scale}
-            // TODO: support showing labels on axes
-            stroke={xAxisTheme.strokeColor}
-            strokeWidth={xAxisTheme.strokeWidth}
-            tickStroke={xAxisTheme.tickStrokeColor}
+            scale={yAxisScale}
+            label={yAxisLabel}
+            labelProps={{
+              fill: yAxisTheme.labelColor,
+              fontSize: yAxisTheme.labelFontSize,
+              textAnchor: yAxisTheme.labelTextAnchor,
+            }}
+            stroke={yAxisTheme.strokeColor}
+            strokeWidth={yAxisTheme.strokeWidth}
+            tickStroke={yAxisTheme.tickStrokeColor}
             // TODO: modify it as a function
             numTicks={getNumberOfTicks(height, data)}
-            tickLabelProps={getXtickLabelProps(xAxisTheme)}
+            tickLabelProps={getXtickLabelProps(yAxisTheme)}
             tickFormat={tickFormat}
             // TODO: format the ticks based on the axis types
             // tickComponent={({ formattedValue, ...tickProps }) => (
@@ -100,17 +129,23 @@ export const AxisLayer: React.SFC<AxisLayerProps> = ({
             // )}
           />
         )}
+        {/* X Axis */}
         {showBottomAxis && (
           <AxisBottom
             top={height}
-            scale={xAxis.scale}
-            stroke={yAxisTheme.strokeColor}
-            strokeWidth={yAxisTheme.strokeWidth}
-            tickStroke={yAxisTheme.tickStrokeColor}
+            scale={xAxisScale}
+            label={xAxisLabel}
+            labelProps={{
+              fill: xAxisTheme.labelColor,
+              fontSize: xAxisTheme.labelFontSize,
+              textAnchor: xAxisTheme.labelTextAnchor,
+            }}
+            stroke={xAxisTheme.strokeColor}
+            strokeWidth={xAxisTheme.strokeWidth}
+            tickStroke={xAxisTheme.tickStrokeColor}
             numTicks={getNumberOfTicks(width, data)}
-            label="label"
             tickFormat={tickFormat}
-            tickLabelProps={getYtickLabelProps(yAxisTheme)}
+            tickLabelProps={getYtickLabelProps(xAxisTheme)}
           />
         )}
       </>