+++ packages/graph/src/themes/index.tsx
@@ -4,6 +4,11 @@ import deepmerge from 'deepmerge';
 
 import { Theme } from '../common/types';
 
+const COLORS = {
+  GRAY: '#7c8a94',
+  DARK: '#233142',
+};
+
 interface ThemeProviderProps {
   /** theme prop can be partial of Theme type */
   theme: Partial<Theme>;
@@ -13,23 +18,42 @@ interface ThemeProviderProps {
 export const themes = {
   default: {
     colors: {
+      /** see [d3-scale-chromatic](https://github.com/d3/d3-scale-chromatic) */
       category: schemeCategory10,
       sequential: {
         scheme: schemeBlues,
         interpolator: interpolateCool,
       },
     },
+    globalStyle: {
+      fontColor: COLORS.DARK,
+    },
     xAxis: {
-      strokeColor: '#7c8a94',
-      tickStrokeColor: '#7c8a94',
+      labelFontSize: 15,
+      labelColor: COLORS.DARK,
+      labelTextAnchor: 'middle',
+      strokeColor: COLORS.GRAY,
+      tickStrokeColor: COLORS.GRAY,
       strokeWidth: 1.5,
       tickFontSize: 13,
     },
     yAxis: {
-      strokeColor: '#7c8a94',
-      tickStrokeColor: '#7c8a94',
+      labelFontSize: 15,
+      labelColor: COLORS.DARK,
+      labelTextAnchor: 'middle',
+      strokeColor: COLORS.GRAY,
+      tickStrokeColor: COLORS.GRAY,
       strokeWidth: 1.5,
       tickFontSize: 13,
+    },
+    headerBox: {
+      titleFontSize: 19,
+      titleColor: COLORS.DARK,
+      titleDescFontSize: 15,
+      titleDescColor: COLORS.GRAY,
+      lineHeight: 1.3,
+      padding: '0 1rem',
+      defaultTextAlign: 'left',
     },
   },
 };