+++ packages/graph/src/utils/styled-components/index.ts
@@ -10,7 +10,6 @@ const {
   css,
   createGlobalStyle,
   keyframes,
-  ThemeProvider,
 } = styledComponents as styledComponents.ThemedStyledComponentsModule<ThemeInterface>;
 
-export { styled, css, createGlobalStyle, keyframes, ThemeProvider };
+export { styled, css, createGlobalStyle, keyframes };